src/PythonProgram/dist_aperture_sim/requirements.txt
numpy>=1.20
scipy>=1.5
opencv-python>=4.5
Pillow>=9.0
PyQt5>=5.15
pydantic>=1.10
PyYAML>=6.0
imagecodecs>=2023.3.16; platform_system != "Windows" and platform_machine != "arm64"
glymur>=0.12.8; platform_system != "Windows" and platform_machine != "arm64"
pytest>=7.0
aiofiles>=23.2
asyncio; python_version>="3.8"
